Add a review comment before merging the branch
